whatpm/t/tree-test-foreign.dat

#data
<math></math>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <math math>

#data
<math>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <math math>

#data
<math><mi></mi></math>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>

#data
<math><mi>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>

#data
<svg></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|     "x"

#data
<svg>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<svg/>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|     "x"

#data
<svg><textpath></textpath></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg textPath>
|     "x"

#data
<svg><foreignobject/></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|     "x"

#data
<svg><foreignObject></foreignobject></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|     "x"

#data
<svg><foreignObject>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>

#data
<svg><foreignobject></body>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         "x"

#data
<p><svg><foreignobject><address>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <address>

#data
<p><svg><foreignobject><article>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <article>

#data
<p><svg><foreignobject><aside>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <aside>

#data
<p><svg><foreignobject><blockquote>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <blockquote>

#data
<p><svg><foreignobject><center>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <center>

#data
<p><svg><foreignobject><datagrid>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <datagrid>

#data
<p><svg><foreignobject><details>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <details>

#data
<p><svg><foreignobject><dialog>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dialog>

#data
<p><svg><foreignobject><dir>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dir>

#data
<p><svg><foreignobject><div>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <div>

#data
<p><svg><foreignobject><dl>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dl>

#data
<p><svg><foreignobject><figure>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <figure>

#data
<p><svg><foreignobject><footer>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <footer>

#data
<p><svg><foreignobject><h6>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <h6>

#data
<p><svg><foreignobject><header>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <header>

#data
<p><svg><foreignobject><menu>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <menu>

#data
<p><svg><foreignobject><nav>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <nav>

#data
<p><svg><foreignobject><ol>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <ol>

#data
<p><svg><foreignobject><p>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <p>

#data
<p><svg><foreignobject><section>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <section>

#data
<p><svg><foreignobject><ul>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <ul>

#data
<p><svg><foreignobject><pre>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <pre>

#data
<p><svg><foreignobject><listing>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <listing>

#data
<p><svg><foreignobject><form>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <form>

#data
<p><svg><foreignobject><fieldset>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <fieldset>

#data
<p><svg><foreignobject><li>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <li>

#data
<p><svg><foreignobject><dd>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dd>

#data
<p><svg><foreignobject><dt>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <dt>

#data
<p><svg><foreignobject><plaintext>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <plaintext>

#data
<address><svg><foreignobject></address>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <address>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<article><svg><foreignobject></article>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<aside><svg><foreignobject></aside>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<ul><svg><foreignobject></ul>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <ul>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<form><svg><foreignobject></form>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <form>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<p><svg><foreignobject></p>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <p>
|           "x"

#data
<dd><svg><foreignobject></dd>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <dd>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<dt><svg><foreignobject></dt>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <dt>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<li><svg><foreignobject></li>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <li>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<h1><svg><foreignobject></h1>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <h1>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<nobr><svg><foreignobject><nobr></nobr></foreignObject></svg></nobr>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <nobr>
|       <svg svg>
|         <svg foreignObject>
|           <nobr>

#data
<button><svg><foreignobject><button></button></foreignObject></svg></button>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <svg svg>
|         <svg foreignObject>
|           <button>

#data
<applet><svg><foreignobject><applet></applet></foreignObject></svg></applet>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <applet>
|       <svg svg>
|         <svg foreignObject>
|           <applet>

#data
<applet><svg><foreignobject></applet>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <applet>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<button><svg><foreignobject></button>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<marquee><svg><foreignobject></marquee>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <marquee>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<object><svg><foreignobject></object>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <object>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<p><svg><foreignobject><table></table>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <table>

#data
<p><svg><foreignobject><hr>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg foreignObject>
|           <hr>

#data
<option><svg><foreignobject></option>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <option>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<optgroup><svg><foreignobject></optgroup>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <optgroup>
|       <svg svg>
|         <svg foreignObject>
|           "x"

#data
<ruby><rt><svg><foreignobject><rp>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rt>
|         <svg svg>
|           <svg foreignObject>
|             <rp>

#data
<ruby><rt><svg><foreignobject><rt>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rt>
|         <svg svg>
|           <svg foreignObject>
|             <rt>

#data
<table><tr><td><svg><caption>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg caption>
|                 "a"

#data
<table><tr><td><svg><col>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg col>
|                 "a"

#data
<table><tr><td><svg><colgroup>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg colgroup>
|                 "a"

#data
<table><tr><td><svg><tr>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg tr>
|                 "a"

#data
<table><caption><svg><caption>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg caption>
|             "a"

#data
<table><caption><svg><col>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg col>
|             "a"

#data
<table><caption><svg><colgroup>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg colgroup>
|             "a"

#data
<table><caption><svg><thead>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg thead>
|             "a"

#data
<table><caption><svg><tr>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg tr>
|             "a"

#data
<table><caption><svg><td>a
#errors
no DOCTYPE
eof in foreign
table not closed
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg td>
|             "a"

#data
<table><tr><td><svg><g>a</td>b
#errors
no DOCTYPE
not closed
char in table
eof
#document
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "a"

#data
<table><tr><th><svg><g>a</th>b
#errors
no DOCTYPE
not closed
char in table
eof
#document
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <svg svg>
|               <svg g>
|                 "a"

#data
<table><caption><svg><g>a</td>b
#errors
no DOCTYPE
unmatched end tag
eof in foreign
eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</th>b
#errors
no DOCTYPE
unmatched end tag
eof in foreign
eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</caption>b
#errors
no DOCTYPE
unmatched end tag
char in table
eof
#document
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "a"

#data
<table><tr><td><svg><g>a</caption>b
#errors
no DOCTYPE
unmatched end tag
eof in foreign
eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><tr><th><svg><g>a</caption>b
#errors
no DOCTYPE
unmatched end tag
eof in foreign
eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <svg svg>
|               <svg g>
|                 "ab"

#data
<table><tr><td><svg><g>a</table>b
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "a"
|     "b"

#data
<table><tr><td><svg><g>a</tbody>b
#errors
no DOCTYPE
unmatched end tag
char in table
no </table>
#document
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "a"

#data
<table><tr><td><svg><g>a</tr>b
#errors
no DOCTYPE
unmatched end tag
char in table
no </table>
#document
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "a"

#data
<table><thead><tr><td><svg><g>a</thead>b
#errors
no DOCTYPE
unmatched end tag
char in table
no </table>
#document
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <thead>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "a"

#data
<table><tfoot><tr><td><svg><g>a</tfoot>b
#errors
no DOCTYPE
unmatched end tag
char in table
no </table>
#document
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "a"

#data
<table><caption><svg><g>a</table>b
#errors
no DOCTYPE
unmatched end tag (implied </caption>)
unmatched end tag </table>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "a"
|     "b"

#data
<table><caption><svg><g>a</tbody>b
#errors
no DOCTYPE
unmatched end tag
eof in foreign
eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</thead>b
#errors
no DOCTYPE
unmatched end tag
eof in foreign
eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</tfoot>b
#errors
no DOCTYPE
unmatched end tag
eof in foreign
eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><caption><svg><g>a</tr>b
#errors
no DOCTYPE
unmatched end tag
eof in foreign
eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "ab"

#data
<table><tr><svg><g>a</tr>b<!---->
#errors
no DOCTYPE
element in <tr>
unmatched end tag
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <tbody>
|         <tr>
|         <!--  -->

#data
<table><tbody><svg><g>a</tbody>b<!---->
#errors
no DOCTYPE
element in table
unmatched end tag
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <tbody>
|       <!--  -->

#data
<table><thead><svg><g>a</thead>b<!---->
#errors
no DOCTYPE
element in table
unmatched end tag
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <thead>
|       <!--  -->

#data
<table><tfoot><svg><g>a</tfoot>b<!---->
#errors
no DOCTYPE
element in table
unmatched end tag
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <tfoot>
|       <!--  -->

#data
<table><svg><g>a</table>b
#errors
no DOCTYPE
element in table
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     <table>
|     "b"

#data
<table><tbody><tr><svg><g>a</tbody>b<!---->
#errors
no DOCTYPE
element in table
unmatched end tag
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <tbody>
|         <tr>
|       <!--  -->

#data
<table><thead><tr><svg><g>a</thead>b<!---->
#errors
no DOCTYPE
element in table
unmatched end tag
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <thead>
|         <tr>
|       <!--  -->

#data
<table><tfoot><tr><svg><g>a</tfoot>b<!---->
#errors
no DOCTYPE
element in table
unmatched end tag
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <tfoot>
|         <tr>
|       <!--  -->

#data
<table><tbody><tr><svg><g>a</table>b
#errors
no DOCTYPE
element in table
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     <table>
|       <tbody>
|         <tr>
|     "b"

#data
<table><tbody><tr><svg><g>a</tbody>b<!---->
#errors
no DOCTYPE
element in table
char in table
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <tbody>
|         <tr>
|       <!--  -->

#data
<table><thead><tr><svg><g>a</thead>b<!---->
#errors
no DOCTYPE
element in table
char in table
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <thead>
|         <tr>
|       <!--  -->

#data
<table><tfoot><tr><svg><g>a</tfoot>b<!---->
#errors
no DOCTYPE
element in table
char in table
eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "a"
|     "b"
|     <table>
|       <tfoot>
|         <tr>
|       <!--  -->

#data
<table><colgroup><svg></colgroup><!----></table>
#errors
no DOCTYPE
element in table
unmatched end tag
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <!--  -->
|     <table>
|       <colgroup>

#data
<!DOCTYPE HTML>
<p xmlns:xlink="http://www.w3.org/1999/xhtml">x</p>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       xmlns:xlink="http://www.w3.org/1999/xhtml"
|       "x"

#data
<!DOCTYPE HTML>
<svg xmlns:xlink="http://www.w3.org/1999/xlink"></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xmlns xlink="http://www.w3.org/1999/xlink"

#data
<!DOCTYPE HTML>
<svg xmlns:xlink="http://www.w3.org/1999/xhtml"></svg>
#errors
bad namespace
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xmlns xlink="http://www.w3.org/1999/xhtml"

#data
<!DOCTYPE HTML>
<svg xmlns:xlink="http://www.w3.org/1999/xlink "></svg>
#errors
bad namespace
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xmlns xlink="http://www.w3.org/1999/xlink "
